Weapon Collider on Sprite and hit timing
The functionality that the Collider2D can be added directly to the sprite in the SpriteEditor as a Custom Physics Shape and does not have to be created as a PolygonCollider2D in the world.

It is also possible to specify the exact frame at which the hit is to be performed and not after half of the animation time as was previously the case.