Replaced the old IntroManager
Replaced the old IntroManager with a new version that is modular and can be changed at any time. It allows for more customization and removes overhead. It creates the intro objects like text, image or video at runtime.
The text can be changed is stored in variables and can be changed later to support multiple languages.