Continue, New Game and Load Game
This one is a bigger update.
I've implemented a new version of the continue button. This continues your latest session as per usual but the loading process is started right after starting the game. So whilst the intro is playing the latest session is already loaded for the players convenience.

Next I've added a very simple new game screen that is not finished by all means but from a coding POV its just the character customization missing that since we're most likely changing the character down the line this would be additional work.

Lastly I've added a loading screen when the user can see all of his saved games with some additional information about the save like player name, money, playtime and last played date. In addition to that there is a load, duplicate and delete button for each save.