Credits plus test
Added the modularity from the intro to the credits.
It's now possible to set up major and minor positions as well as images. Those are scrolled by on the screen until everyone and every image has been seen.
The user has the ability to fast forward or skip the credits completely.´

I've also added some test credits for testers to review.