Basic NPC Behaviour
This update adds all kinds o basic functionality to create a NPC and have a NPC do its thing.

In detail I've added:
NPCDefinition: This defines a single character e.g. Kira, it also defines a home and working location as well as character traits, name etc.
NPCSchedule & NPCScheduleEntry: This defines the daily routine a character is going to follow. This is strict with a location the character has to move to and when he will leave that location and what kind of activity it is.
This is run by the NPCScheduleRunner.
The NPC moves with the NPCMovementController and interacts with the NPCInteractionController.

NPCs can even have a quick chat on the go when they see an other NPC or the player.

This update is only the bare minimum to have NPCs in the game and has a lot of potential!