Advanced Animal Behaviour
Added a Animal Controller that controls the components of a animal.
Added a state machine for the animal basically desicidng what the animal is doing with its time like idle, walk, eat, sleep, etc.
Added a animal visual that shows the name of the animal and later its state like fed, pet and produced item.
Added a animal story where animals can be bought. This is a extent of the store class that will be refactored in the future with other stores.